backend_new/pyproject.toml
[project]
name = "backend-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.9",
    "django-cors-headers>=4.4.0",
    "drf-nested-routers>=0.93.4",
    "django-filter>=25.2",
    "djangorestframework-simplejwt>=5.3.1",
]
